Created the Dashboard and create links on the dashboard
