+++ pages/dashboard.jsx
@@ -1,0 +1,9 @@
+const Dashboard = () => {
+  return (
+    <div>
+      <h1>Dashboard</h1>
+    </div>
+  );
+};
+
+export default Dashboard;
