+++ pages/user/link/create.jsx
@@ -1,0 +1,9 @@
+const Create = () => {
+  return (
+    <div>
+      <h1>Create</h1>
+    </div>
+  );
+};
+
+export default Create;
