+++ src/components/atom/images.jsx
