+++ src/utils/links.jsx
@@ -1,0 +1,7 @@
+import Link from "next/link";
+
+export const links = (href, children) => (
+  <Link href={href} passHref>
+    {children}
+  </Link>
+);
